// admin-dashboard/src/app/services/models/driver-profile-view.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DriverProfileView {
  createdAt?: string;
  firstName?: string;
  lastName?: string;
  licenceDocument?: string;
  licenceNumber?: string;
  nationalId?: string;
  phone?: string;
  preferredLanes?: string;
  profilePic?: string;
  publicId?: string;
  remark?: string;
  status?: string;
  updatedAt?: string;
}