// admin-dashboard/src/app/services/models/page-response-driver-profile-view.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DriverProfileView } from '../models/driver-profile-view';
export interface PageResponseDriverProfileView {
  content?: Array<DriverProfileView>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}